KA-183 Add GET method unit test, implement IEquatable
Quite big refactoring to the model had to be done in order
to make the controller testable. The Equals method had to be overriden and because of that, the GetHashCode as well.

However, to make GetHashCode work as intended, Id needed to be readonly, which proved to be quite difficult to achieve. The XML formatter requires a propperty to have public setter method.

Eventually, I settled up for implementing an empty public setter and private readonly "_id" field. See comments for further explanation.

With regards to this change, a minor changes to the list items initializations had to be made in the tests and in the controller itself.