KA-183 Add custom action filters and Guid binder
The goal of this commit was to remove validation logic from the controller methods.

The argument validation has been moved out from the controller into custom action filters. The three filters now take care of:
- Null arguments
- Model validity
- Consistency of GUIDs in Put request

All validation has been removed from the controller, the filters have been added in WebApiConfig (with exception of the GUID consistency filter, as it only matters to PUT request)

A custom GUID model binder has been created, as the default one was not correctly returning an error on invalid guid formats (a "null" value was passed to the ModelState dictionary - this did consequently cause an error, as the key of this value was not nullable; however, the resulting error message was not informative enough)

This model binder was then applied globally to all Guid type parameters in the WebApiConfig. To reflect on this change, all types of Id arguments were changed from String to Guid, and all checks whether the argument can be parsed were removed from the controller - the binder takes care of this functionality now.

All unit tests that were using string guid arguments will not work in this commit and need to be refactored to reflect the change!

Some comments were added to clarify code, some unused "using" statements were removed.