KA-183 Refactor the filters to make testable, add unit tests
The filters require injectable HttpRequestMessage, as it is used to generate the error response in case the filter fails. An abstract class with injectable request message was created to base the filters on; each filter now overrides the DoValidation abstract method to its own needs and gets the request object passed into it.